+++ moped-editor/src/views/projects/projectsListView/helpers.js
@@ -7,7 +7,7 @@ import { formatDateType, formatTimeStampTZType } from "src/utils/dateAndTime";
 import Pagination from "../../../components/GridTable/Pagination";
 import ExternalLink from "../../../components/ExternalLink";
 import ProjectStatusBadge from "../projectView/ProjectStatusBadge";
-import RenderSignalLink from "../signalProjectTable/RenderSignalLink";
+import RenderSignalLink from "../../../components/RenderSignalLink";
 
 export const filterNullValues = (value) => {
   if (!value || value === "-") {
